+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorBaseDerivedRequiredTest.kt
@@ -27,7 +27,9 @@ package net.pwall.json.schema.codegen
 
 import kotlin.test.Test
 import kotlin.test.expect
+
 import java.io.File
+
 import net.pwall.json.JSON
 import net.pwall.json.pointer.JSONPointer
 import net.pwall.json.schema.codegen.CodeGeneratorTestUtil.OutputDetails
@@ -47,6 +49,18 @@ class CodeGeneratorBaseDerivedRequiredTest {
         codeGenerator.generateAll(JSON.parse(input), JSONPointer("/\$defs"))
         expect(createHeader("Base.kt") + expectedBase) { outputDetailsBase.output() }
         expect(createHeader("Derived.kt") + expectedDerived) { outputDetailsDerived.output() }
+    }
+
+    @Test fun `should generate base and derived classes where derived adds required in Java`() {
+        val input = File("src/test/resources/test-base-derived-required.schema.json")
+        val codeGenerator = CodeGenerator(TargetLanguage.JAVA)
+        val outputDetailsBase = OutputDetails(TargetFileName("Base", "java", dirs))
+        val outputDetailsDerived = OutputDetails(TargetFileName("Derived", "java", dirs))
+        codeGenerator.basePackageName = "com.example"
+        codeGenerator.outputResolver = outputCapture(outputDetailsBase, outputDetailsDerived)
+        codeGenerator.generateAll(JSON.parse(input), JSONPointer("/\$defs"))
+        expect(createHeader("Base.java") + expectedBaseJava) { outputDetailsBase.output() }
+        expect(createHeader("Derived.java") + expectedDerivedJava) { outputDetailsDerived.output() }
     }
 
     companion object {
@@ -103,6 +117,131 @@ class Derived(
 
 }
 """
+
+        const val expectedBaseJava =
+"""package com.example;
+
+public class Base {
+
+    private final String aaa;
+
+    public Base(
+            String aaa
+    ) {
+        this.aaa = aaa;
+    }
+
+    public String getAaa() {
+        return aaa;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (this == other)
+            return true;
+        if (!(other instanceof Base))
+            return false;
+        Base typedOther = (Base)other;
+        return aaa == null ? typedOther.aaa == null : aaa.equals(typedOther.aaa);
+    }
+
+    @Override
+    public int hashCode() {
+        return (aaa != null ? aaa.hashCode() : 0);
+    }
+
+    public static class Builder {
+
+        private String aaa;
+
+        public Builder withAaa(String aaa) {
+            this.aaa = aaa;
+            return this;
+        }
+
+        public Base build() {
+            return new Base(
+                    aaa
+            );
+        }
+
+    }
+
+}
+"""
+
+        const val expectedDerivedJava =
+"""package com.example;
+
+public class Derived extends Base {
+
+    private final String bbb;
+
+    public Derived(
+            String aaa,
+            String bbb
+    ) {
+        super(validateAaa(aaa));
+        if (bbb == null)
+            throw new IllegalArgumentException("Must not be null - bbb");
+        this.bbb = bbb;
+    }
+
+    private static String validateAaa(String aaa) {
+        if (aaa == null)
+            throw new IllegalArgumentException("Must not be null - aaa");
+        return aaa;
+    }
+
+    public String getBbb() {
+        return bbb;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (this == other)
+            return true;
+        if (!(other instanceof Derived))
+            return false;
+        if (!super.equals(other))
+            return false;
+        Derived typedOther = (Derived)other;
+        return bbb.equals(typedOther.bbb);
+    }
+
+    @Override
+    public int hashCode() {
+        int hash = super.hashCode();
+        return hash ^ bbb.hashCode();
+    }
+
+    public static class Builder {
+
+        private String aaa;
+        private String bbb;
+
+        public Builder withAaa(String aaa) {
+            this.aaa = aaa;
+            return this;
+        }
+
+        public Builder withBbb(String bbb) {
+            this.bbb = bbb;
+            return this;
+        }
+
+        public Derived build() {
+            return new Derived(
+                    aaa,
+                    bbb
+            );
+        }
+
+    }
+
+}
+"""
+
     }
 
 }